+++ CHANGELOG.md
@@ -13,7 +13,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 - 重构为monorepo 并添加 zentao-api 包 ([e540a6a](https://github.com/aceHubert/zentao/commit/e540a6acf925764dc879f805ef24d946139727cc))
 - **zentao-api:** 重构文档结构并移除通用客户端 ([f7ffe47](https://github.com/aceHubert/zentao/commit/f7ffe47b9acfc710a704ee5a3f6bd2275a8e0d8d))
-- **zentao-clients:** 集成 @ace-zentao/api 统一封装各版本客户端 ([51a7736](https://github.com/aceHubert/zentao/commit/51a77367c0eddcbd1ec4c908efd6ce9c42e516ed))
+- **zentao-clients:** 集成 @acehubert/zentao-api 统一封装各版本客户端 ([51a7736](https://github.com/aceHubert/zentao/commit/51a77367c0eddcbd1ec4c908efd6ce9c42e516ed))
 - **zentao-mcp:** 新增独立 CLI 工具并优化 MCP 启动参数 ([cfe9270](https://github.com/aceHubert/zentao/commit/cfe92700c8ca6883bf47e025cb3c5c1b02c692d8))
 
 # Changelog